[smbscan][spider.py] Scanner output change.
New output treats paths and directories separately. Paths have full names and
no total size and items. Directories have short names and both size and items
count.
Formally, new format is the following:

1 <path_id> <path>
0 <path_id> <sharedir_id> <size> <self_path_id> <items> <filename>

where path_id in the second line means path_id for file's parent and
self_path_id means path_id for diercory itself or zero for file
Items is the number of files and subdirectories in the directory or zero if
the line is for a file

For each directory the scanner prints line of first type when it first
encounters that directory and line of second type is printed when scanner
leaves the directory after processing all it's contents recursively